requirement.txt
addict==2.4.0
antlr4-python3-runtime==4.9.3
asciitree==0.3.3
asttokens==2.4.1
astunparse==1.6.3
attrdict==2.0.1
autobahn==23.1.2
autolab-core==1.1.1
automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.3
cachetools==4.2.4
chardet==5.2.0
cloudpickle==3.0.0
colorlog==6.8.0
comm==0.2.2
configargparse==1.7
constantly==23.10.4
contourpy==1.1.1
cycler==0.12.1
cython==3.0.7
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
diffusers==0.11.1
dill==0.3.7
docker-pycreds==0.4.0
egl-probe==1.0.2
einops==0.7.0
einops-exts==0.0.4
embreex==2.17.7.post4
executing==2.0.1
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.13.1
flask==3.0.2
fonttools==4.47.2
freetype-py==2.4.0
future==0.18.3
gast==0.3.3
gdown==5.1.0
gitdb==4.0.11
gitpython==3.1.40
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gym==0.23.1
gym-notices==0.0.8
h5py==2.10.0
huggingface-hub==0.23.0
hydra==2.5
hydra-core==1.3.2
hyperlink==21.0.0
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-resources==6.1.1
incremental==22.10.0
ipython==8.12.3
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-core==5.7.2
jupyterlab-widgets==3.0.10
keras-preprocessing==1.1.2
kiwisolver==1.4.5
lazy-loader==0.3
llvmlite==0.41.1
lxml==4.9.4
mapbox-earcut==1.0.1
matplotlib==3.7.4
matplotlib-inline==0.1.6
meshcat==0.3.2
msgpack==1.0.7
multiprocess==0.70.15
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==2.2
ninja==1.11.1.1
numba==0.58.1
numcodecs==0.12.1
numpy-quaternion==2023.0.2
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pkgutil-resolve-name==1.3.10
plotly==5.20.0
plyfile==1.0.3
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pybullet==3.2.6
pycollada==0.6
pygame==2.1.0
pygccxml==2.2.1
pyglet==1.5.0
pygments==2.17.2
pyngrok==7.1.6
pyopengl==3.1.0
pyparsing==3.1.1
pyplusplus==1.8.5
pyquaternion==0.9.9
pyrender==0.1.45
pysdf==0.1.9
python-dateutil==2.8.2
python-fcl==0.7.0.5
pytorch-lightning==1.9.5
pytz==2024.1
pyvirtualdisplay==3.0
pywavelets==1.4.1
pyzmq==26.0.3
ray==2.9.0
referencing==0.32.0
regex==2024.4.28
retrying==1.3.4
roslibpy==1.6.0
robomimic==0.3.0
rotary-embedding-torch==0.2.1
rpds-py==0.15.2
rtree==1.1.0
ruamel-yaml==0.18.5
ruamel-yaml-clib==0.2.8
scikit-learn==1.3.2
scipy==1.8.0
sentry-sdk==1.39.1
service-identity==24.1.0
setproctitle==1.3.3
setuptools-scm==8.0.4
shapely==2.0.2
shortuuid==1.0.11
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
svg-path==6.3
swig==4.1.1.post1
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tomli==2.0.1
torchtyping==0.1.4
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
trimesh==3.23.5
twisted==24.3.0
txaio==23.1.1
typeguard==4.1.5
tzdata==2023.4
u-msgpack-python==2.8.0
urdfpy==0.0.22
wandb==0.12.21
warp-lang==0.10.1
wcwidth==0.2.13
widgetsnbextension==4.0.10
wrapt==1.16.0
xvfbwrapper==0.2.9
xxhash==3.4.1
yacs==0.1.8
rl-games==1.6.1
yourdfpy==0.0.53
zope-interface==6.2
